Move date helpers to a separate file for reuse
